requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-core
python-dotenv
arxiv
wikipedia
duckduckgo-search
ddgs
pydantic
requests
